Switched from allochronic::select to futures_util
Allochronic is causing high CPU usage due to its implementation of
select not ever returning Poll::Pending. The fix is more complicated
than just switching to futures_util for now.

futures_util requires streams to be fused, so I've added the extra logic
to ReceiverStream so that we don't need to wrap the future in a Fuse --
allowing stop() to still be called.