Added Transmog compatability.
This commit adds support for specifying the serialization format at the
time of accepting a connection, which means that ALPN protocol
negotiation can be used to control which serialization format is used on
an incoming connection.

The next feature is the ability to switch serialization formats on a
per-stream basis, after the r#type negotation.