Attempting to get quinn upgrade working
This commit attempts to upgrade quinn, but unfortunately many unit tests
are failing. I'm not sure if the tests are valid given the updated quinn
or if the updated code isn't compatible for some reason.

The biggest thorn in this upgrade is that Quinn no longer implements
Stream for some of its types, hence the manual pinning and looping done
in some locations. I believe the way these are written is fine, since
the select macros are going to drop the future in the situation of a
shutdown anyways, so there's no worry about aborting an in-progress
future and then losing state by asking for a new incoming connection.