Restoring closing of incoming connections
After the last commit, I wasn't able to figure out how to get new
connections to be rejected. I asked, and while there may be a better way
to do this in the future, providing a new ServerConfig with the
concurrent connections set to 0 blocks new connections from being
formed.

Thank you to both @djc and @Ralith for the help on this!