Fixing stream refactor
When refactoring away from streams, I thought I could simplify the
select_biased logic. I thought I could do this because the once the task
is shut down, it should be able to be dropped. I'm not sure which part
of this did not work as expected, but by properly looping and handling
the 'complete' flow, everything works as intended.

The other change that was needed, however, is to update Endpoint's close
to also explicitly close incoming. Without this, the task is never shut
down since calling close() doesn't abort the pending Accept future.