Hand-implementing incoming "stream" futures
As much as I wanted to make select_biased work, I just couldn't. The
example was hanging with the previous implementation, and my attempts at
fixing it were fruitless without also breaking the unit tests.

I decided to try to hand rolling the futures, and magically, everything
worked. Because Accept isn't Unpin, it's a little ugly and requires a
Box::pin currently. I don't love having that allocation there, but I
couldn't figure out how to get the right access through pin_project, and
I know this approach is 100% safe. It probably can be done safely
without the allocation, however, as the only time that the accept field
needs to be overwritten is after the future has been polled -- thus the
Pin is not longer being used.

I feel like this logic should have been able to be representable with
select/select_biased, which might avoid the issue entirely, as pin!
seemingly was able to deal Accept being Unpin, since it was being stored
on the stack. However, I failed to get the logic to get both the example
*and* the unit test working at the same time.